requirements.txt
aiofiles==24.1.0
html5tagger==1.3.0
httptools==0.6.4
multidict==6.4.3
PyMySQL==1.1.1
PyYAML==6.0.2
sanic==25.3.0
sanic-routing==23.12.0
setuptools==80.4.0
tracerite==1.1.1
typing_extensions==4.13.2
websockets==15.0.1
